pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ascii-art-mcp"
version = "0.1.0"
description = "MCP server for converting images to ASCII art with photo and logo modes"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "tikket" },
]
keywords = ["mcp", "ascii-art", "image-processing", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
]

[project.optional-dependencies]
hifi = ["scikit-image>=0.19.0"]

[project.scripts]
ascii-art-mcp = "ascii_art_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ascii_art_mcp"]
